Revert "Updated all test case parameters"
This reverts commit a6d640294d233c7b5ccecd38cb36cfee9a3ff951.

+++ src/test/java/seedu/planitarium/person/FamilyTest.java
@@ -21,8 +21,8 @@ public class FamilyTest {
     private static final boolean SILENT = true;
     private static final boolean NOT_SILENT = false;
     private static final int FIRST_ENTRY = 1;
-    private static final Double VALID_AMOUNT = 1000.0;
-    private static final Double NEW_AMOUNT = 500.0;
+    private static final int VALID_AMOUNT = 1000;
+    private static final int NEW_AMOUNT = 500;
     private static final String VALID_DESCRIPTION = "Testing";
     private static final boolean PERMANENT = true;
